+++ rust/test/all_test.go
@@ -13,7 +13,6 @@ import (
 	"time"
 
 	"github.com/streamingfast/dstore"
-
 	pbsubstreams "github.com/streamingfast/substreams/pb/sf/substreams/v1"
 	"github.com/streamingfast/substreams/state"
 	"github.com/streamingfast/substreams/wasm"