+++ CHANGELOG.md
@@ -7,6 +7,7 @@ project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+* Added `substreams::log::info!(...)` and `substreams::log::debug!(...)` supporting formatting arguments (acts like `println!()` macro).
 * Added new field `logs_truncated` that can be used to determined if logs were truncated.
 * Augmented logs truncation limit to 128 KiB per module per block.
 * Updated `substreams run` to properly report module progress error.